@@ -1,6 +1,6 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '2.7.14'
+	id 'org.springframework.boot' version '3.2.3'
 	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
 }
 
@@ -8,7 +8,7 @@ group = 'company'
 version = '0.0.1-SNAPSHOT'
 
 java {
-	sourceCompatibility = '11'
+	sourceCompatibility = '17'
 }
 
 configurations {
@@ -39,8 +39,11 @@ dependencies {
 
 	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
 
-	implementation 'javax.servlet:jstl'
 	implementation 'org.webjars:bootstrap:5.1.3'
+	compileOnly group: 'jakarta.servlet', name: 'jakarta.servlet-api', version: '6.0.0'
+	implementation group: 'jakarta.transaction', name: 'jakarta.transaction-api', version: '2.0.1'
+	implementation group: 'jakarta.persistence', name: 'jakarta.persistence-api', version: '3.1.0'
+	implementation group: 'jakarta.validation', name: 'jakarta.validation-api', version: '3.0.2'
 	implementation 'org.webjars:webjars-locator:0.45'
 	implementation 'org.apache.tomcat.embed:tomcat-embed-jasper'
 	implementation ("org.springframework.boot:spring-boot-starter-oauth2-client")
