@@ -8,6 +8,7 @@ import company.board_project.domain.list.leaguelist.service.LeagueListService;
 import company.board_project.domain.match.leaguematch.entity.LeagueMatch;
 import company.board_project.global.response.MultiResponseDto;
 import company.board_project.domain.team.service.TeamService;
+import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import lombok.extern.slf4j.Slf4j;
@@ -17,7 +18,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.constraints.Positive;
 import java.util.List;
 
 @RestController
