@@ -7,6 +7,7 @@ import company.board_project.domain.schedule.dto.ScheduleResponseDto;
 import company.board_project.domain.schedule.entity.Schedule;
 import company.board_project.domain.schedule.mapper.ScheduleMapper;
 import company.board_project.domain.schedule.service.ScheduleService;
+import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -14,7 +15,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.constraints.Positive;
 import java.util.List;
 
 @RestController
