@@ -8,13 +8,13 @@ import company.board_project.domain.team.entity.Team;
 import company.board_project.domain.team.repository.TeamRepository;
 import company.board_project.domain.user.entity.User;
 import company.board_project.domain.user.service.UserService;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import javax.transaction.Transactional;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
