@@ -5,16 +5,15 @@ import company.board_project.global.security.auth.dto.AuthEmailDto;
 import company.board_project.global.security.auth.dto.AuthEmailQuestionDto;
 import company.board_project.global.security.auth.dto.AuthPasswordDto;
 import company.board_project.global.security.auth.service.AuthService;
+import jakarta.mail.MessagingException;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-
-import javax.mail.MessagingException;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.Valid;
-import javax.validation.constraints.NotBlank;
 
 @RestController
 @RequestMapping("/auth")
