@@ -1,9 +1,8 @@
 package company.board_project.global.security.auth.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
 
 @NoArgsConstructor
 @Getter
