@@ -2,10 +2,10 @@ package company.board_project.global.security.utils;
 import com.google.gson.Gson;
 import company.board_project.global.exception.BusinessLogicException;
 import company.board_project.global.exception.ErrorResponse;
+import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 public class ErrorResponder {
